Implemeneted pipe process fix that Martim figured out. Divided functions (for norminette) Organized functions in respective files (Can be ajusted).
