+++ src/exec/exec_utils.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exe.c                                              :+:      :+:    :+:   */
+/*   exec_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 11:22:20 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/09/27 12:59:18 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/10/03 13:32:22 by ddiniz-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,22 +79,4 @@ void	built_ins(char *builtin)
 		;/* env() */
 	if (ft_strcmp(builtin, "exit") == 0)
 		;/* exit() */
-}
-
-int	exec(t_cmdlist *cmdlist, char **paths, char **envp)
-{
-	char		*cmd_path;
-
-	cmd_path = NULL;
-	if (!is_built_in(cmdlist->content->cmd_flags[0]))
-		cmd_path = is_exec(cmdlist->content->cmd_flags[0], paths);
-	else
-	{
-		built_ins(cmdlist->content->cmd_flags[0]);
-		exit (0);
-	}
-	if (cmd_path && execve(cmd_path, cmdlist->content->cmd_flags, envp) == -1)
-		perror("EXECVE ERROR\n");
-	free(cmd_path);
-	return (0);
 }
