+++ src/exec/open_file.c
@@ -6,7 +6,7 @@
 /*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/27 12:03:12 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/09/28 11:37:53 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/10/03 13:25:26 by ddiniz-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
