+++ src/exec/redir_hdoc.c
@@ -6,7 +6,7 @@
 /*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/15 13:43:52 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/09/28 14:17:40 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/10/03 13:32:01 by ddiniz-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ int	redir_in_out(t_content *content, char **arr, int pos)
 
 	tmp = content;
 	size = arr_size(arr);
-	while (pos < size && ft_strcmp(arr[pos], "|") != 0 )
+	while (pos < size && ft_strcmp(arr[pos], "|") != 0)
 	{
 		if (tmp->input && ft_strcmp(arr[pos], "<") == 0)
 			open_file_in(tmp, tmp->input);
@@ -61,103 +61,10 @@ int	redir_in_out(t_content *content, char **arr, int pos)
 	return (0);
 }
 
-int	no_pipe(t_minishell *ms, char **paths, char **envp, char **arr)
+void	set_fd(t_minishell *ms)
 {
-	t_cmdlist	*tmp;
-	pid_t		child;
-	
-	(void)arr;
-	tmp = ms->cmdlist;
-	child = fork();
-	if (child == 0)
-	{
-		//redir_in_out(tmp->content, ms->main_arr, 0);
-		exec(tmp, paths, envp);
-	}
-	else
-	{
-		wait(NULL);
-		dup2(ms->fdout_buf, STDOUT_FILENO);
-		close(ms->fdout_buf);
-		dup2(ms->fdin_buf, STDIN_FILENO);
-		close(ms->fdin_buf);
-	}
-	return (0);
-}
-
-int	ft_pipe(t_minishell *ms, t_cmdlist *cmdlist, int *pipe_fd, char **envp, char **paths, int j)
-{
-	pid_t		child;
-	
-	if (pipe(pipe_fd) < 0)
-	{
-		printf("Pipe error\n");
-		return (1);
-	}
-	child = fork();
-	if (child == -1)
-		return (printf("Fork Error\n"));
-	if (child == 0)
-	{
-		if (redir_check(ms))
-		{
-			redir_in_out(cmdlist->content, ms->main_arr, j);
-		}
-		else
-		{	
-			close(pipe_fd[0]);
-			dup2(pipe_fd[1], STDOUT_FILENO);
-			close(pipe_fd[1]);
-		}
-		exec(cmdlist, paths, envp);
-	}
-	else
-	{
-		close(pipe_fd[1]);
-		wait (NULL);
-		dup2(pipe_fd[0], STDIN_FILENO);
-		close(pipe_fd[0]);
-	}
-	return (1);
-}
-
-int	run(t_minishell *ms, char **envp)
-{
-	int			i;
-	int			j;
-	int			cmds;
-	t_cmdlist	*tmp;
-	char		**paths;
-	int			pipe_fd[2];
-
-	i = 0;
-	j = 0;
-	tmp = ms->cmdlist;
-	paths = path_init(envp);
-	cmds = ms->cmd_count;
-	if (cmds == 1) //if there isn't a pipe
-	{
-		redir_in_out(tmp->content, ms->main_arr, 0);
-		no_pipe(ms, paths, envp, ms->main_arr);
-		free_array(paths);
-		return (1);
-	}
-	while (cmds > 1 && i < cmds) //if there's a pipe
-	{
-		ft_pipe(ms, tmp, pipe_fd, envp, paths, j);
-		while (ms->main_arr[j] && ft_strcmp(ms->main_arr[j], "|") != 0)
-			j++;
-		if (tmp->next)
-			tmp = tmp->next;
-		j++;
-		i++;
-	}
-	if (tmp && !redir_check(ms))
-		exec(tmp, paths, envp);
 	dup2(ms->fdout_buf, STDOUT_FILENO);
 	close(ms->fdout_buf);
 	dup2(ms->fdin_buf, STDIN_FILENO);
 	close(ms->fdin_buf);
-	free_array(paths);
-	return (0);
 }
